src/core/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
boto3==1.29.3
pydantic==2.4.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1 